Refactor module and permission setup for console and uc
