+++ ui/console-src/main.ts
@@ -1,19 +1,15 @@
 import { setLanguage, setupI18n } from "@/locales";
 import { setupApiClient } from "@/setup/setupApiClient";
 import { setupComponents } from "@/setup/setupComponents";
+import { setupCoreModules, setupPluginModules } from "@/setup/setupModules";
 import "@/setup/setupStyles";
+import { setupUserPermissions } from "@/setup/setupUserPermissions";
 import { setupVueQuery } from "@/setup/setupVueQuery";
-import { useRoleStore } from "@/stores/role";
-import {
-  setupCoreModules,
-  setupPluginModules,
-} from "@console/setup/setupModules";
+import modules from "@console/modules";
 import { useThemeStore } from "@console/stores/theme";
-import { consoleApiClient } from "@halo-dev/api-client";
-import { stores, utils } from "@halo-dev/ui-shared";
+import { stores } from "@halo-dev/ui-shared";
 import "core-js/es/object/has-own";
 import { createPinia } from "pinia";
-import type { DirectiveBinding } from "vue";
 import { createApp } from "vue";
 import App from "./App.vue";
 import router from "./router";
@@ -27,46 +23,16 @@ setupApiClient();
 
 app.use(createPinia());
 
-async function loadUserPermissions() {
-  const { data: currentPermissions } =
-    await consoleApiClient.user.getPermissions({
-      name: "-",
-    });
-  const roleStore = useRoleStore();
-  roleStore.$patch({
-    permissions: currentPermissions,
-  });
-
-  // Set permissions in shared utils
-  utils.permission.setUserPermissions(currentPermissions.uiPermissions);
-
-  app.directive(
-    "permission",
-    (el: HTMLElement, binding: DirectiveBinding<string[]>) => {
-      const { value } = binding;
-      const { any } = binding.modifiers;
-
-      if (utils.permission.has(value, any ?? false)) {
-        return;
-      }
-
-      el?.remove?.();
-    }
-  );
-}
-
 async function loadActivatedTheme() {
   const themeStore = useThemeStore();
   await themeStore.fetchActivatedTheme();
 }
 
-(async function () {
-  await initApp();
-})();
+await initApp();
 
 async function initApp() {
   try {
-    setupCoreModules(app);
+    setupCoreModules({ app, router, platform: "console", modules });
 
     const currentUserStore = stores.currentUser();
     await currentUserStore.fetchCurrentUser();
@@ -80,17 +46,15 @@ async function initApp() {
       return;
     }
 
-    await loadUserPermissions();
+    await setupUserPermissions(app);
 
     try {
-      await setupPluginModules(app);
+      await setupPluginModules({ app, router, platform: "console" });
     } catch (e) {
       console.error("Failed to load plugins", e);
     }
 
-    if (globalInfoStore.globalInfo?.userInitialized) {
-      await loadActivatedTheme();
-    }
+    await loadActivatedTheme();
   } catch (e) {
     console.error(e);
   } finally {
