+++ ui/console-src/modules/index.ts
@@ -1,10 +1,8 @@
 import type { PluginModule } from "@halo-dev/ui-shared";
 
-const modules = Object.values(
-  import.meta.glob("./**/module.ts", {
-    eager: true,
-    import: "default",
-  })
-) as PluginModule[];
+const modules = import.meta.glob("./**/module.ts", {
+  eager: true,
+  import: "default",
+}) as Record<string, PluginModule>;
 
 export default modules;
